agents/MCP-Client/requirements.txt
anyio>=4.6.2.post1
asyncio>=3.4.3
openai>=1.55.3
python-dotenv>=1.0.1
requests>=2.32.3
rich>=13.9.4
httpx